requirements.txt
bleach
dependency-injector
fastapi
fastapi-limiter
httpx
loguru
motor
pydantic
python-dotenv
redis
uvicorn

